Semaine1/projet/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
fastapi==0.115.0
ffmpy==0.4.0
fonttools==4.54.1
fsspec==2024.9.0
gradio==4.44.1
gradio_client==1.3.0
huggingface-hub==0.25.2
imageio==2.35.1
importlib_resources==6.4.5
kiwisolver==1.4.7
lazy_loader==0.4
matplotlib==3.9.2
networkx==3.3
numpy==2.1.2
opencv-python==4.10.0.84
orjson==3.10.7
pandas==2.2.3
pillow==10.4.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyparsing==3.1.4
python-multipart==0.0.12
pytz==2024.2
ruff==0.6.9
scikit-image==0.24.0
scipy==1.14.1
semantic-version==2.10.0
setuptools==75.1.0
shellingham==1.5.4
starlette==0.38.6
tifffile==2024.9.20
tomlkit==0.12.0
tqdm==4.66.5
typer==0.12.5
tzdata==2024.2
uvicorn==0.31.0
websockets==12.0
